requirements.txt
azure-core==1.35.0
azure-storage-blob==12.26.0
azure-storage-file-datalake==12.21.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.7
et_xmlfile==2.0.0
greenlet==3.2.4
idna==3.10
isodate==0.7.2
numpy==2.2.6
openpyxl==3.1.5
pandas==2.3.2
pycparser==2.22
PyMySQL==1.1.2
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
